requirements.txt
-e .
streamlit>=1.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
numpy>=1.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
deepgram-sdk>=3.0.0
streamlit-mic-recorder>=0.0.8
beautifulsoup4>=4.12.0
requests>=2.31.0
deep-translator>=1.11.0
mysql-connector-python>=8.0.0
google-generativeai>=0.8.0
groq>=0.9.0
fpdf2>=2.7.0
